Better way to run integration test, integrate the server app into the test progress, so no need to start a standalong server progress
